requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
sqlalchemy==2.0.36
databases==0.9.0
pydantic==2.10.0
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
python-multipart==0.0.19
PyJWT>=2.0.0
pg8000==1.31.2
cloud-sql-python-connector[pg8000]
pandas
numpy